Add missing app tests
