+++ apps/presheaves-and-opposite-categories/src/App.test.tsx
@@ -1,0 +1,9 @@
+import { render, screen } from '@testing-library/react';
+import { expect, test } from 'vitest';
+import App from './App';
+
+test('renders heading', () => {
+  render(<App />);
+  const heading = screen.getByText(/Presheaf Visualization/i);
+  expect(heading).toBeDefined();
+});
